requirements.txt
arabic-reshaper==2.1.1
argcomplete==1.12.3
argon2-cffi==20.1.0
asgiref==3.5.0
astroid==2.4.2
attrs==21.2.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
bleach==4.1.0
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
cycler==0.11.0
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
distlib==0.3.3
dj-database-url==0.5.0
Django==3.2.12
django-autoslug==1.9.8
django-ckeditor==6.0.0
django-crispy-forms==1.10.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-multiselectfield==0.1.12
django-rest-knox==4.2.0
django-rest-passwordreset==1.2.1
django-rest-swagger==2.2.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
docopt==0.6.2
entrypoints==0.3
Faker==8.12.1
filelock==3.3.1
future==0.18.2
gunicorn==20.0.4
html5lib==1.1
idna==2.10
importlib-metadata==4.7.0
ipykernel==6.2.0
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.7.0
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.1
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.2
lazy-object-proxy==1.4.3
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
mccabe==0.6.1
mistune==0.8.4
mysqlclient==2.0.3
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.3
numpy==1.21.1
openapi-codec==1.3.2
packaging==21.0
pandas==1.3.4
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.0.1
pipreqs==0.4.11
platformdirs==2.4.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
psycopg2==2.8.6
pycparser==2.20
Pygments==2.10.0
PyJWT==1.7.1
pylint==2.6.0
PyMySQL==1.0.2
pyparsing==2.4.7
PyPDF2==1.26.0
pypi==2.1
pyrsistent==0.18.0
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2020.4
pywin32==301
pywinpty==1.1.3
PyYAML==6.0
pyzmq==22.2.1
qtconsole==5.1.1
QtPy==1.10.0
razorpay==1.2.0
reportlab==3.5.56
requests==2.25.1
Send2Trash==1.8.0
simplejson==3.17.6
six==1.15.0
sqlparse==0.4.1
terminado==0.11.1
testpath==0.5.0
text-unidecode==1.3
tk==0.1.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
twilio==6.51.1
typed-ast==1.4.2
typing-extensions==3.10.0.0
uritemplate==4.1.1
urllib3==1.26.2
virtualenv==20.8.1
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.1.0
widgetsnbextension==3.5.1
wrapt==1.12.1
xhtml2pdf==0.2.5
yarg==0.1.9
zipp==3.5.0
